Move SQLiteProvider into outer layout.tsx to fix drizzle studio
